django/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
distlib==0.3.3
Django==3.2.7
filelock==3.2.0
mysql-connector-python==8.0.26
mysqlclient==2.0.3
platformdirs==2.4.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
virtualenv==20.8.1
python-decouple==3.5 
requests==2.26.0 
drf-jwt==1.19.1
djangorestframework==3.12.4
django-cors-headers==3.10.0
google-api-python-client==2.31.0              
google-auth==2.3.3               
google-auth-httplib2==0.1.0               
google-auth-oauthlib==0.4.6               
googleapis-common-protos==1.53.0
oauth2client==4.1.3
PyDrive2==1.10.0
xhtml2pdf==0.2.5
PyPDF2==1.26.0